CA-239854 : Remove the need to access hal.dll
We used to access hal.dll to find out precisely which shutdown
command to use.  We don't need that in 2k8r2 and above.

We have, however, continued to acces hal.dll to output the version
in xenstore /os/attr/hal

There is no good reason to do this - and I can't believe anyone
uses it.  So I'm going to remove this and see if anyone complains.

(We have, on odd situations,r un into problems accessing hal.dll
they are rare - but better to remove a moving part that doesn't
need to move, than to let it bite us somewhere down the road)