CP-24040: VDA Backchannel API - Handle Pipe Disconnects
If the named pipe to XenConsole Monitor is disconnected, dispose it
and inform all subscribers of the 'PipeDisconnected' event.

Connection to the named pipe now happens explicitly when calling
'Start()', and not in the Constructor.

An exception is thrown, if 'Start()' is called and the
'PipeDisconnected' event has no subscribers.

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>