CA-264903 : All messages for the console dll are bounded by 0x02,0x03
The console DLL thinks in terms of messages
But xencons thinks in terms of bytes

This extends the NamedPipeClientStreamWrapper to make a translation
from bytes to messages.

This also converts the NamesPipeClientStreamStub into a PipeStream
which makes it easier to test as a drop in replacement for a
NamedPipeClientStream

Finally tests are added, and fixed put in place to make testing
work better in this new world.